assets: update_mirror.sh: Correctly parse dist value from snap
snap_name and dist fails to be parsed if DISTS includes
distribution names including hyphen (e.g. buster-updates).
This results in that only snapshot of buster-updates is always
published as debian/buster (snapshot of buster is overwritten)
when we use the default settings for Debian repository:

  REPO=debian
  OS_RELEASE=buster
  DISTS=( ${OS_RELEASE} ${OS_RELEASE}-updates )

The correct dist value can be parse from snap on the assumption
that snap values always start with REPO.

Signed-off-by: Kazuhiro Hayashi <zuka0828@gmail.com>